+++ Sources/el-grocer-shopper-sdk-iOS/SourceFiles/Modules/Orders/OrderHistoryCell/CandCHistoryCell/CandCHistoryCell.swift
@@ -161,7 +161,7 @@ class CandCHistoryCell: UITableViewCell {
         guard status != nil else {
             return
         }
-        let progress : Float = status!.stepNumber.floatValue / totalStep
+        let progress : Float = status!.stepNumber / totalStep
         self.progressView.setProgress(progress , animated: true)
     }
